Deprecate a bunch of FileHandlers functions
