+++ spimdisasm/mips/FilesHandlers.py
@@ -1,11 +1,10 @@
 #!/usr/bin/env python3
 
-# SPDX-FileCopyrightText: © 2022 Decompollaborate
+# SPDX-FileCopyrightText: © 2022-2023 Decompollaborate
 # SPDX-License-Identifier: MIT
 
 from __future__ import annotations
 
-import dataclasses
 from typing import TextIO
 from pathlib import Path
 
@@ -15,6 +14,7 @@ from .. import common
 
 from . import sections
 from . import symbols
+from . import FunctionRodataEntry
 
 
 def createSectionFromSplitEntry(splitEntry: common.FileSplitEntry, array_of_bytes: bytearray, outputPath: Path, context: common.Context) -> sections.SectionBase:
@@ -59,39 +59,9 @@ def writeSection(path: Path, fileSection: sections.SectionBase):
     return path
 
 
-@dataclasses.dataclass
-class FunctionRodataEntry:
-    function: symbols.SymbolFunction | None = None
-    rodataSyms: list[symbols.SymbolBase] = dataclasses.field(default_factory=list)
-    lateRodataSyms: list[symbols.SymbolBase] = dataclasses.field(default_factory=list)
-
-    @staticmethod
-    def getEntryFromSections(func: symbols.SymbolFunction, rodataSection: sections.SectionRodata) -> FunctionRodataEntry:
-        rodataList: list[symbols.SymbolBase] = []
-        lateRodataList: list[symbols.SymbolBase] = []
-
-        intersection = func.instrAnalyzer.referencedVrams & rodataSection.symbolsVRams
-        if len(intersection) == 0:
-            return FunctionRodataEntry(func)
-
-        for rodataSym in rodataSection.symbolList:
-            if rodataSym.vram not in intersection:
-                continue
-
-            if not rodataSym.shouldMigrate():
-                continue
-
-            if rodataSym.contextSym.isLateRodata():
-                lateRodataList.append(rodataSym)
-            else:
-                rodataList.append(rodataSym)
-
-        return FunctionRodataEntry(func, rodataList, lateRodataList)
-
-
 #! @deprecated
 def getRdataAndLateRodataForFunctionFromSection(func: symbols.SymbolFunction, rodataSection: sections.SectionRodata) -> tuple[list[symbols.SymbolBase], list[symbols.SymbolBase], int]:
-    entry = FunctionRodataEntry.getEntryFromSections(func, rodataSection)
+    entry = FunctionRodataEntry.getEntryFromSection(func, rodataSection)
 
     lateRodataSize = 0
     for rodataSym in entry.lateRodataSyms:
@@ -99,69 +69,29 @@ def getRdataAndLateRodataForFunctionFromSection(func: symbols.SymbolFunction, ro
 
     return entry.rodataSyms, entry.lateRodataSyms, lateRodataSize
 
+#! @deprecated
 def getRdataAndLateRodataForFunction(func: symbols.SymbolFunction, rodataFileList: list[sections.SectionBase]) -> tuple[list[symbols.SymbolBase], list[symbols.SymbolBase], int]:
-    rdataList: list[symbols.SymbolBase] = []
-    lateRodataList: list[symbols.SymbolBase] = []
+    entry = FunctionRodataEntry.getEntryFromPossibleRodataSections(func, rodataFileList)
+
     lateRodataSize = 0
+    for rodataSym in entry.lateRodataSyms:
+        lateRodataSize += rodataSym.sizew
 
-    for rodataSection in rodataFileList:
-        assert isinstance(rodataSection, sections.SectionRodata)
+    return entry.rodataSyms, entry.lateRodataSyms, lateRodataSize
 
-        if len(rdataList) > 0 or len(lateRodataList) > 0:
-            # We already have the rodata for this function. Stop searching
-            break
-
-        # Skip the file if there's nothing in this file refenced by the current function
-        intersection = func.instrAnalyzer.referencedVrams & rodataSection.symbolsVRams
-        if len(intersection) == 0:
-            continue
-
-        rdataList, lateRodataList, lateRodataSize = getRdataAndLateRodataForFunctionFromSection(func, rodataSection)
-
-    return rdataList, lateRodataList, lateRodataSize
-
-# Note: the `lateRodataSize` parameter is deprecated and ignored, it has been kept because of compatibility reasons
+#! @deprecated
 def writeFunctionRodataToFile(f: TextIO, func: symbols.SymbolFunction, rdataList: list[symbols.SymbolBase], lateRodataList: list[symbols.SymbolBase], lateRodataSize: int=0):
+    entry = FunctionRodataEntry(func, rdataList, lateRodataList)
+    entry.writeToFile(f, writeFunction=False)
 
-    if len(rdataList) > 0:
-        # Write the rdata
-        sectionName = ".rodata"
-        f.write(f".section {sectionName}" + common.GlobalConfig.LINE_ENDS)
-        for sym in rdataList:
-            f.write(sym.disassemble(migrate=True, useGlobalLabel=True))
-            f.write(common.GlobalConfig.LINE_ENDS)
-
-    if len(lateRodataList) > 0:
-        # Write the late_rodata
-        f.write(".section .late_rodata" + common.GlobalConfig.LINE_ENDS)
-
-        lateRodataSize = 0
-        for sym in lateRodataList:
-            lateRodataSize += sym.sizew
-
-        if lateRodataSize / len(func.instructions) > 1/3:
-            align = 4
-            firstLateRodataVram = lateRodataList[0].vram
-            if firstLateRodataVram is not None and firstLateRodataVram % 8 == 0:
-                align = 8
-            f.write(f".late_rodata_alignment {align}" + common.GlobalConfig.LINE_ENDS)
-        for sym in lateRodataList:
-            f.write(sym.disassemble(migrate=True, useGlobalLabel=True))
-            f.write(common.GlobalConfig.LINE_ENDS)
-
-    if len(rdataList) > 0 or len(lateRodataList) > 0:
-        f.write(common.GlobalConfig.LINE_ENDS + ".section .text" + common.GlobalConfig.LINE_ENDS)
-
+#! @deprecated
 def writeSplitedFunction(path: Path, func: symbols.SymbolFunction, rodataFileList: list[sections.SectionBase]):
     path.mkdir(parents=True, exist_ok=True)
+    entry = FunctionRodataEntry.getEntryFromPossibleRodataSections(func, rodataFileList)
 
     funcPath = path / (func.getName()+ ".s")
     with funcPath.open("w") as f:
-        rdataList, lateRodataList, lateRodataSize = getRdataAndLateRodataForFunction(func, rodataFileList)
-        writeFunctionRodataToFile(f, func, rdataList, lateRodataList, lateRodataSize)
-
-        # Write the function itself
-        f.write(func.disassemble(migrate=True))
+        entry.writeToFile(f, writeFunction=True)
 
 def writeOtherRodata(path: Path, rodataFileList: list[sections.SectionBase]):
     for rodataSection in rodataFileList:
@@ -178,43 +108,6 @@ def writeOtherRodata(path: Path, rodataFileList: list[sections.SectionBase]):
             with rodataSymbolPath.open("w") as f:
                 f.write(".section .rodata" + common.GlobalConfig.LINE_ENDS)
                 f.write(rodataSym.disassemble(migrate=True))
-
-
-def getFunctionsAndRodataSymsFromSections(textSection: sections.SectionText, rodataSection: sections.SectionRodata) -> list[FunctionRodataEntry]:
-    allUnmigratedRodataSymbols: list[symbols.SymbolBase] = []
-
-    for rodataSym in rodataSection.symbolList:
-        if not rodataSym.shouldMigrate():
-            # We only care for the symbols which will not be migrated
-            allUnmigratedRodataSymbols.append(rodataSym)
-
-    allEntries: list[FunctionRodataEntry] = []
-
-    for func in textSection.symbolList:
-        assert isinstance(func, symbols.SymbolFunction)
-
-        entry = FunctionRodataEntry.getEntryFromSections(func, rodataSection)
-
-        if len(entry.rodataSyms) > 0:
-            firstFuncRodataSym = entry.rodataSyms[0]
-
-            while len(allUnmigratedRodataSymbols) > 0:
-                rodataSym = allUnmigratedRodataSymbols[0]
-
-                if rodataSym.vram >= firstFuncRodataSym.vram:
-                    # Take all the symbols up to the first rodata sym referenced by the current function
-                    break
-
-                allEntries.append(FunctionRodataEntry(rodataSyms=[rodataSym]))
-                del allUnmigratedRodataSymbols[0]
-
-        allEntries.append(entry)
-
-    # Check if there's any rodata symbol remaining and add it to the list
-    for rodataSym in allUnmigratedRodataSymbols:
-        allEntries.append(FunctionRodataEntry(rodataSyms=[rodataSym]))
-
-    return allEntries
 
 
 def writeMigratedFunctionsList(processedSegments: dict[common.FileSectionType, list[sections.SectionBase]], functionMigrationPath: Path, name: str) -> None:
